refactor: Answers QnAService에 적용

+++ src/main/java/nextstep/qna/service/QnAService.java
@@ -30,12 +30,8 @@ public class QnAService {
             throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
         }
 
-        List<Answer> answers = question.getAnswers();
-        for (Answer answer : answers) {
-            if (!answer.isOwner(loginUser)) {
-                throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
-            }
-        }
+        final Answers answers = new Answers(question.getAnswers());
+        answers.validateAnswersOwnership(loginUser);
 
         final DeleteHistories deleteHistories = new DeleteHistories();
         deleteHistories.add(question);
